Presentation/src/app/components/game/game.component.ts
import { Component, OnInit } from '@angular/core';
import { filter, map, Observable } from 'rxjs';
import { HandShape } from 'src/app/shared/interfaces/hand-shape.interface';
import { GameService } from '../../services/game.service';
import { utils } from '../../shared/util/utils';

@Component({
  selector: 'app-game',
  templateUrl: './game.component.html',
  styles: [
  ]
})
export class GameComponent implements OnInit {

  handShapes$: Observable<HandShape[]> = this.gameService.evtRestResponse$.pipe(
    filter(response => response && response.data),
    map(response => response.data)
  )

  errors$ = this.gameService.evtRestResponse$.pipe(
    filter(response => utils.isNotNullNorUndefined(response)),
    map(response => utils.isNotNullNorUndefined(response.error))
  )

  constructor(private readonly gameService: GameService) { }

  ngOnInit(): void {
    this.gameService.getHandShapes();
  }

}
